requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.12.0
tiktoken==0.5.2
python-dotenv==1.0.0
jupyter==1.0.0
pandas==2.2.0
matplotlib==3.8.2
requests==2.31.0
beautifulsoup4==4.12.3
